Refactors academic calendar module and entities
Removes deprecated AcademicCalendar entity and module
Adds new AcademicCalendar entity with updated fields and relationships
Fixes import paths in academic-calendar.module.ts

Relates to #123
