+++ backend/src/school/school.entity.ts
@@ -12,7 +12,7 @@ import {
 import { AdmissionPolicy } from '../admission-policy/admission-policy.entity';
 import { GradingSystem } from '../grading-system/grading-system.entity';
 import { FeeStructure } from '../fee-structure/fee-structure.entity';
-import { AcademicCalendar } from '../academic-calendar/academic-calendar.entity';
+import { AcademicCalendar } from '../academic-calendar/entitites/academic-calendar.entity';
 import { Profile } from 'src/profile/profile.entity';
 import { SchoolAdmin } from 'src/school-admin/school-admin.entity';
 import { Student } from 'src/student/student.entity';
